requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
model-mommy==2.0.0
pytz==2023.3rst
sqlparse==0.4.4
typing_extensions==4.7.0
